Documents/GitHub/A2A-Instabids/requirements.txt
# Core A2A Server & Client Dependencies
fastapi
uvicorn[standard]
httpx
pydantic

# Google Agent Development Kit (Assuming it's installed separately or add path dependency if needed)
# google-adk

# Supabase Client
supabase

# Google Cloud Services (Vision, potentially others later)
google-cloud-vision

# Environment Variables
python-dotenv

# Logging
logging
